% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monkeytest-package.R
\docType{package}
\name{monkeytest-package}
\alias{monkeytest}
\alias{monkeytest-package}
\title{monkeytest: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EmilHvitfeldt/monkeytest}
  \item \url{https://emilhvitfeldt.github.io/monkeytest/}
  \item Report bugs at \url{https://github.com/EmilHvitfeldt/monkeytest/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emilhhvitfeldt@gmail.com} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})

}
\keyword{internal}
