% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure.R
\name{test_file_pure}
\alias{test_file_pure}
\title{Run all tests in a single file purely}
\usage{
test_file_pure(path, ...)
}
\arguments{
\item{path}{Path to file.}

\item{...}{Additional parameters passed on to \code{testthat::test_file()}}
}
\description{
Helper, setup, and teardown files located in the same directory as the
test will also be run. These pure functions will only run \code{test_that()} calls.
}
\examples{
path <- testthat::testthat_example("success")
test_file_pure(path)
test_file_pure(path, reporter = "minimal")
}
