% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure-package.R
\name{test_package_pure}
\alias{test_package_pure}
\alias{test_check_pure}
\alias{test_local_pure}
\title{Run all tests in a package purely}
\usage{
test_package_pure(package, reporter = check_reporter(), ...)

test_check_pure(package, reporter = check_reporter(), ...)

test_local_pure(path = ".", reporter = NULL, ...)
}
\arguments{
\item{package}{If these tests belong to a package, the name of the package.}

\item{reporter}{Reporter to use to summarise output. Can be supplied
as a string (e.g. "summary") or as an R6 object
(e.g. \code{SummaryReporter$new()}).

See \link[testthat]{Reporter} for more details and a list of built-in reporters.}

\item{...}{Additional arguments passed to \code{\link[testthat:test_dir]{testthat::test_dir()}}}

\item{path}{Path to directory containing tests.}
}
\value{
A list (invisibly) containing data about the test results.
}
\description{
These pure functions will only run \code{test_that()} calls.
\itemize{
\item \code{test_local_pure()} tests a local source package.
\item \code{test_package_pure()} tests an installed package.
\item \code{test_check_pure()} checks a package during \verb{R CMD check}.
}

Tests live in \code{tests/testthat}.
}
\section{Special files}{


There are two types of \code{.R} file that have special behaviour:
\itemize{
\item Test files start with \code{test} and are executed in alphabetical order.
\item Setup files start with \code{setup} and are executed before tests. If
clean up is needed after all tests have been run, you can use
\code{withr::defer(clean_up(), teardown_env())}. See \code{vignette("test-fixtures")}
for more details.
}

There are two other types of special file that we no longer recommend using:
\itemize{
\item Helper files start with \code{helper} and are executed before tests are
run. They're also loaded by \code{devtools::load_all()}, so there's no
real point to them and you should just put your helper code in \verb{R/}.
\item Teardown files start with \code{teardown} and are executed after the tests
are run. Now we recommend interleave setup and cleanup code in \verb{setup-}
files, making it easier to check that you automatically clean up every
mess that you make.
}

All other files are ignored by testthat.

}

\section{Environments}{


Each test is run in a clean environment to keep tests as isolated as
possible. For package tests, that environment that inherits from the
package's namespace environment, so that tests can access internal functions
and objects.

}

